# back/docs/swagger.yaml
definitions:
  account.AccountCreateDto:
    properties:
      email:
        maxLength: 320
        minLength: 6
        type: string
      language:
        allOf:
        - $ref: '#/definitions/constants.AccountLanguage'
        enum:
        - en
        - fr
      password:
        maxLength: 100
        type: string
      termsAccepted:
        type: boolean
      termsVersion:
        type: string
      timeZone:
        type: string
    required:
    - email
    - language
    - password
    - termsAccepted
    - termsVersion
    - timeZone
    type: object
  account.AccountProviderDto:
    properties:
      provider:
        $ref: '#/definitions/constants.Provider'
    type: object
  account.AccountResponseDto:
    properties:
      avatarUrl:
        type: string
      color:
        type: string
      email:
        type: string
      language:
        $ref: '#/definitions/constants.AccountLanguage'
      providers:
        items:
          $ref: '#/definitions/account.AccountProviderDto'
        type: array
      timeZone:
        type: string
      userName:
        type: string
    type: object
  account.AccountUpdateDto:
    properties:
      color:
        type: string
      email:
        maxLength: 320
        minLength: 6
        type: string
      language:
        allOf:
        - $ref: '#/definitions/constants.AccountLanguage'
        enum:
        - en
        - fr
      password:
        maxLength: 100
        type: string
      termsAccepted:
        type: boolean
      termsVersion:
        type: string
      timeZone:
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    type: object
  account.ForgotPasswordDto:
    properties:
      email:
        maxLength: 320
        minLength: 6
        type: string
    required:
    - email
    type: object
  account.ResetPasswordDto:
    properties:
      password:
        maxLength: 100
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  availability.AvailabilityCreateDto:
    properties:
      endsAt:
        type: string
      startsAt:
        type: string
    required:
    - endsAt
    - startsAt
    type: object
  availability.AvailabilityResponseDto:
    properties:
      endsAt:
        type: string
      id:
        type: string
      startsAt:
        type: string
    type: object
  availability.AvailabilityUpdateDto:
    properties:
      endsAt:
        type: string
      startsAt:
        type: string
    type: object
  constants.AccountLanguage:
    enum:
    - en
    - fr
    type: string
    x-enum-varnames:
    - ACCOUNT_LANGUAGE_EN
    - ACCOUNT_LANGUAGE_FR
...
  constants.Provider:
    enum:
    - google
    - discord
    - github
    type: string
    x-enum-varnames:
    - PROVIDER_GOOGLE
    - PROVIDER_DISCORD
    - PROVIDER_GITHUB
...
  helpers.ApiError:
    properties:
      code:
        type: string
    type: object
...
  slot.ConfirmSlotDto:
    properties:
      endsAt:
        type: string
      startsAt:
        type: string
    required:
    - endsAt
    - startsAt
    type: object
  slot.SlotResponseDto:
    properties:
      endsAt:
        type: string
      id:
        type: string
      isValidated:
        type: boolean
      startsAt:
        type: string
    type: object
...
info:
  contact:
    email: contact@zide.fr
  description: SlotFinder API Doc
  title: SlotFinder API
  version: 1.0b
paths:
  /api/v1/account:
    patch:
      consumes:
      - application/json
      description: Update own account
      parameters:
      - description: Account parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.AccountUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.AccountResponseDto'
        "400":
          description: 'Bad Request - Code can be: ERR_USERNAME_ALREADY_TAKEN, ERR_INVALID_PASSWORD_FORMAT,
            or ERR_INVALID_COLOR_FORMAT'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      security:
      - AccessTokenCookie: []
      summary: Update my account
      tags:
      - Account
    post:
      consumes:
      - application/json
      description: Create a new account with the provided parameters.
      parameters:
      - description: Account parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.AccountCreateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Bad Request - Code can be: ERR_INVALID_EMAIL_FORMAT, ERR_INVALID_PASSWORD_FORMAT,
            or ERR_EMAIL_ALREADY_EXISTS'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      summary: Create an account
      tags:
      - Account
  /api/v1/account/avatar:
    patch:
      consumes:
      - multipart/form-data
      description: UploadAvatar the avatar image of the current user
      parameters:
      - description: Avatar image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ApiError'
      security:
      - AccessTokenCookie: []
      summary: Upload Avatar
      tags:
      - Account
  /api/v1/account/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset email to the user
      parameters:
      - description: Email for password reset
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.ForgotPasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Bad Request - Code can be: ERR_PASSWORD_RESET_TOO_FREQUENT'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      summary: Forgot Password
      tags:
      - Account
  /api/v1/account/me:
    get:
      consumes:
      - application/json
      description: Get the account information of the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.AccountResponseDto'
      security:
      - AccessTokenCookie: []
      summary: Get My Account
      tags:
      - Account
  /api/v1/account/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.ResetPasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Bad Request - Code can be: ERR_INVALID_PASSWORD_FORMAT, ERR_INVALID_RESET_TOKEN,
            or ERR_RESET_TOKEN_EXPIRED'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      summary: Reset Password
      tags:
      - Account
...
  /api/v1/availabilities/{availabilityId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Availability ID
        in: path
        name: availabilityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Bad Request - Code can be: ERR_AVAILABILITY_NOT_FOUND, ERR_AVAILABILITY_ACCESS_DENIED,
            ERR_EVENT_ACCESS_DENIED, or ERR_EVENT_ENDED'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      security:
      - BearerAuth: []
      summary: Delete an availability
      tags:
      - Availability
    patch:
      consumes:
      - application/json
      parameters:
      - description: Availability ID
        in: path
        name: availabilityId
        required: true
        type: string
      - description: Availability parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/availability.AvailabilityUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/availability.AvailabilityResponseDto'
        "400":
          description: 'Bad Request - Code can be: ERR_AVAILABILITY_NOT_FOUND, ERR_AVAILABILITY_ACCESS_DENIED,
            ERR_EVENT_NOT_FOUND, ERR_EVENT_ENDED, ERR_EVENT_ACCESS_DENIED, ERR_EVENT_START_AFTER_END,
            ERR_AVAILABILITY_DURATION_TOO_SHORT, ERR_AVAILABILITY_INVALID_TIME_INTERVAL,
            ERR_AVAILABILITY_START_BEFORE_EVENT, or ERR_AVAILABILITY_END_AFTER_EVENT'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      security:
      - BearerAuth: []
      summary: Update an availability
      tags:
      - Availability
...
  /api/v1/events/{eventId}/availability:
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Availability parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/availability.AvailabilityCreateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/availability.AvailabilityResponseDto'
        "400":
          description: 'Bad Request - Code can be: ERR_EVENT_NOT_FOUND, ERR_EVENT_ENDED,
            ERR_EVENT_ACCESS_DENIED, ERR_EVENT_START_AFTER_END, ERR_AVAILABILITY_DURATION_TOO_SHORT,
            ERR_AVAILABILITY_INVALID_TIME_INTERVAL, ERR_AVAILABILITY_START_BEFORE_EVENT,
            or ERR_AVAILABILITY_END_AFTER_EVENT'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      security:
      - BearerAuth: []
      summary: Create an availability
      tags:
      - Availability
...
  /api/v1/slots/{slotId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Slot Id
        in: path
        name: slotId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Bad Request - Code can be: ERR_SLOT_NOT_FOUND or ERR_EVENT_ACCESS_DENIED'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      security:
      - BearerAuth: []
      summary: Remove a validated slot
      tags:
      - Slot
  /api/v1/slots/{slotId}/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: Slot Id
        in: path
        name: slotId
        required: true
        type: string
      - description: Confirm Slot parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/slot.ConfirmSlotDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/slot.SlotResponseDto'
        "400":
          description: 'Bad Request - Code can be: ERR_SLOT_NOT_FOUND, ERR_EVENT_ACCESS_DENIED,
            ERR_EVENT_ENDED, ERR_SLOT_INVALID_STARTS_AT, or ERR_SLOT_INVALID_ENDS_AT'
          schema:
            $ref: '#/definitions/helpers.ApiError'
      security:
      - BearerAuth: []
      summary: Confirm a slot
      tags:
      - Slot
...
securityDefinitions:
  AccessTokenCookie:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"