# back/pkg/availability/availability.controller.go
package availability

import (
	"app/commons/constants"
	"app/commons/guard"
	"app/commons/helpers"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type AvailabilityController struct {
	availabilityService *AvailabilityService
}

func NewAvailabilityController(ctl *AvailabilityController) *AvailabilityController {
	if ctl != nil {
		return ctl
	}

	return &AvailabilityController{
		availabilityService: NewAvailabilityService(nil),
	}
}

// extracts and validates the eventId parameter from the URL path.
func (ctl *AvailabilityController) getEventIdParam(c *gin.Context) (eventIdUuid uuid.UUID, err error) {
	eventId := c.Param("eventId")
	if eventId == "" {
		return eventIdUuid, constants.ERR_EVENT_NOT_FOUND.Err
	}

	eventIdUuid, err = uuid.Parse(eventId)
	if err != nil || eventIdUuid == uuid.Nil {
		return eventIdUuid, constants.ERR_EVENT_NOT_FOUND.Err
	}

	return eventIdUuid, nil
}

// extracts and validates the availability parameter from the URL path.
func (ctl *AvailabilityController) getAvailabilityIdParam(c *gin.Context) (availabilityIdUuid uuid.UUID, err error) {
	availability := c.Param("availabilityId")
	if availability == "" {
		return availabilityIdUuid, constants.ERR_AVAILABILITY_NOT_FOUND.Err
	}

	availabilityIdUuid, err = uuid.Parse(availability)
	if err != nil || availabilityIdUuid == uuid.Nil {
		return availabilityIdUuid, constants.ERR_AVAILABILITY_NOT_FOUND.Err
	}

	return availabilityIdUuid, nil
}

// @Summary Create an availability
// @Tags Availability
// @Accept json
// @Produce json
// @Param eventId path string true "Event ID"
// @Param data body AvailabilityCreateDto true "Availability parameters"
// @Security BearerAuth
// @Success 200 {object} AvailabilityResponseDto
// @Failure 400 {object} helpers.ApiError "Bad Request - Code can be: ERR_EVENT_NOT_FOUND, ERR_EVENT_ENDED, ERR_EVENT_ACCESS_DENIED, ERR_EVENT_START_AFTER_END, ERR_AVAILABILITY_DURATION_TOO_SHORT, ERR_AVAILABILITY_INVALID_TIME_INTERVAL, ERR_AVAILABILITY_START_BEFORE_EVENT, or ERR_AVAILABILITY_END_AFTER_EVENT"
// @Router /api/v1/events/{eventId}/availability [post]
func (ctl *AvailabilityController) Create(c *gin.Context) {
	var data AvailabilityCreateDto
	if err := helpers.SetHttpContextBody(c, &data); err != nil {
		return
	}

	var user *guard.Claims
	if err := guard.GetUserClaims(c, &user); err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	eventId, err := ctl.getEventIdParam(c)
	if err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	availability, err := ctl.availabilityService.Create(&data, eventId, user)

	helpers.HandleJSONResponse(c, availability, err)
}

// @Summary Update an availability
// @Tags Availability
// @Accept json
// @Produce json
// @Param availabilityId path string true "Availability ID"
// @Param data body AvailabilityUpdateDto true "Availability parameters"
// @Security BearerAuth
// @Success 200 {object} AvailabilityResponseDto
// @Failure 400 {object} helpers.ApiError "Bad Request - Code can be: ERR_AVAILABILITY_NOT_FOUND, ERR_AVAILABILITY_ACCESS_DENIED, ERR_EVENT_NOT_FOUND, ERR_EVENT_ENDED, ERR_EVENT_ACCESS_DENIED, ERR_EVENT_START_AFTER_END, ERR_AVAILABILITY_DURATION_TOO_SHORT, ERR_AVAILABILITY_INVALID_TIME_INTERVAL, ERR_AVAILABILITY_START_BEFORE_EVENT, or ERR_AVAILABILITY_END_AFTER_EVENT"
// @Router /api/v1/availabilities/{availabilityId} [patch]
func (ctl *AvailabilityController) Update(c *gin.Context) {
	var data AvailabilityUpdateDto
	if err := helpers.SetHttpContextBody(c, &data); err != nil {
		return
	}

	var user *guard.Claims
	if err := guard.GetUserClaims(c, &user); err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	availabilityId, err := ctl.getAvailabilityIdParam(c)
	if err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	availability, err := ctl.availabilityService.Update(&data, availabilityId, user)

	helpers.HandleJSONResponse(c, availability, err)
}

// @Summary Delete an availability
// @Tags Availability
// @Accept json
// @Produce json
// @Param availabilityId path string true "Availability ID"
// @Security BearerAuth
// @Success 200
// @Failure 400 {object} helpers.ApiError "Bad Request - Code can be: ERR_AVAILABILITY_NOT_FOUND, ERR_AVAILABILITY_ACCESS_DENIED, ERR_EVENT_ACCESS_DENIED, or ERR_EVENT_ENDED"
// @Router /api/v1/availabilities/{availabilityId} [delete]
func (ctl *AvailabilityController) Delete(c *gin.Context) {
	var user *guard.Claims
	if err := guard.GetUserClaims(c, &user); err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	availabilityId, err := ctl.getAvailabilityIdParam(c)
	if err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	err = ctl.availabilityService.Delete(availabilityId, user)

	helpers.HandleJSONResponse(c, nil, err)
}