# back/pkg/slot/slot.controller.go
package slot

import (
	"app/commons/constants"
	"app/commons/guard"
	"app/commons/helpers"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type SlotController struct {
	slotService *SlotService
}

func NewSlotController(ctl *SlotController) *SlotController {
	if ctl != nil {
		return ctl
	}

	return &SlotController{
		slotService: NewSlotService(nil),
	}
}

// @Summary Confirm a slot
// @Tags Slot
// @Param slotId path string true "Slot Id"
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param data body ConfirmSlotDto true "Confirm Slot parameters"
// @Success 200 {object} SlotResponseDto
// @Failure 400 {object} helpers.ApiError "Bad Request - Code can be: ERR_SLOT_NOT_FOUND, ERR_EVENT_ACCESS_DENIED, ERR_EVENT_ENDED, ERR_SLOT_INVALID_STARTS_AT, or ERR_SLOT_INVALID_ENDS_AT"
// @Router /api/v1/slots/{slotId}/confirm [post]
func (ctl *SlotController) ConfirmSlot(c *gin.Context) {
	var user *guard.Claims
	if err := guard.GetUserClaims(c, &user); err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	var data ConfirmSlotDto
	if err := helpers.SetHttpContextBody(c, &data); err != nil {
		return
	}

	slotId, err := uuid.Parse(c.Param("slotId"))
	if err != nil {
		helpers.HandleJSONResponse(c, nil, constants.ERR_SLOT_NOT_FOUND.Err)
		return
	}

	slots, err := ctl.slotService.ConfirmSlot(data, slotId, user.Id)
	helpers.HandleJSONResponse(c, slots, err)
}

// @Summary Remove a validated slot
// @Tags Slot
// @Param slotId path string true "Slot Id"
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200
// @Failure 400 {object} helpers.ApiError "Bad Request - Code can be: ERR_SLOT_NOT_FOUND or ERR_EVENT_ACCESS_DENIED"
// @Router /api/v1/slots/{slotId} [delete]
func (ctl *SlotController) RemoveValidatedSlot(c *gin.Context) {
	var user *guard.Claims
	if err := guard.GetUserClaims(c, &user); err != nil {
		helpers.HandleJSONResponse(c, nil, err)
		return
	}

	slotId, err := uuid.Parse(c.Param("slotId"))
	if err != nil {
		helpers.HandleJSONResponse(c, nil, constants.ERR_SLOT_NOT_FOUND.Err)
		return
	}

	err = ctl.slotService.RemoveValidatedSlot(slotId, user.Id)
	helpers.HandleJSONResponse(c, nil, err)
}